+++ library/src/twiddles.cpp
@@ -21,123 +21,311 @@
 *******************************************************************************/
 
 #include "twiddles.h"
+#include "arithmetic.h"
+#include "device/kernels/twiddle_factors.h"
 #include "function_pool.h"
 #include "rocfft_hip.h"
+#include <cassert>
+#include <math.h>
+#include <numeric>
+#include <stdexcept>
+#include <string>
+#include <tuple>
 
 // Twiddle factors table
 template <typename T>
 class TwiddleTable
 {
-    size_t N; // length
+protected:
+    size_t N;
+    size_t half_N;
     size_t length_limit; // limit number of generated table elements
-
-public:
-    TwiddleTable(size_t _N, size_t _length_limit)
-        : N(_N)
-        , length_limit(_length_limit ? _length_limit : _N)
-    {
-    }
-
-    std::vector<T> GenerateTwiddleTable(const std::vector<size_t>& radices)
-    {
-        // cosine, sine arrays. T is float2 or double2, wc.x stores cosine,
-        // wc.y stores sine
-        std::vector<T> wc(length_limit);
-        const double   TWO_PI = -6.283185307179586476925286766559;
-
-        // Make sure the radices vector multiplication product up to N
-        assert(N
-               == std::accumulate(std::begin(radices),
-                                  std::end(radices),
-                                  static_cast<size_t>(1),
-                                  std::multiplies<>()));
-
-        // Generate the table
-        size_t L  = radices.front();
-        size_t nt = 0;
-        for(auto radix = radices.begin() + 1; radix != radices.end(); ++radix)
-        {
-            L *= *radix;
-
-            // Twiddle factors
-            for(size_t k = 0; k < (L / *radix) && nt < length_limit; k++)
-            {
-                double theta = TWO_PI * (k) / (L);
-
-                for(size_t j = 1; j < *radix && nt < length_limit; j++)
-                {
-                    double c = cos((j)*theta);
-                    double s = sin((j)*theta);
-
-                    // if (fabs(c) < 1.0E-12)    c = 0.0;
-                    // if (fabs(s) < 1.0E-12)    s = 0.0;
-
-                    wc[nt].x = c;
-                    wc[nt].y = s;
-                    nt++;
-                }
-            }
-        } // end of for radices
-        wc.resize(nt);
-
-        return wc;
-    }
-
-    std::vector<T> GenerateTwiddleTable()
-    {
-        // cosine, sine arrays. T is float2 or double2, wc.x stores cosine,
-        // wc.y stores sine
-        std::vector<T> wc(length_limit);
-        const double   TWO_PI = -6.283185307179586476925286766559;
-
-        // Generate the table
-        size_t nt = 0;
-        for(size_t i = 0; i < N && nt < length_limit; i++)
-        {
-            double c = cos(TWO_PI * i / N);
-            double s = sin(TWO_PI * i / N);
-
-            wc[nt].x = c;
-            wc[nt].y = s;
-            nt++;
-        }
-
-        return wc;
-    }
 
     // Attach half-N table for potential fused even-length
     // real2complex post-processing or complex2real pre-processing.
     // Pre/post processing requires a table that's a quarter of the
     // real length, but N here is our complex length.  So half-N is
     // what we need.
-    void AttachHalfNTable(const std::vector<T>& twtc, gpubuf& twts)
+    bool attach_halfN;
+
+    void GetKernelParams(const std::vector<size_t>& radices,
+                         gpubuf_t<size_t>&          radices_device,
+                         gpubuf_t<size_t>&          radices_prod_device,
+                         gpubuf_t<size_t>&          radices_sum_prod_device,
+                         size_t&                    max_radix_prod,
+                         size_t&                    min_radix,
+                         size_t&                    table_sz)
     {
-        size_t         half_N       = (N + 1) / 2;
-        size_t         total_length = twtc.size() + half_N;
-        std::vector<T> twc_all(total_length);
-        std::copy(twtc.begin(), twtc.end(), twc_all.begin());
+        std::vector<size_t> radices_sum_prod;
+        std::vector<size_t> radices_prod;
 
-        const double TWO_PI = -6.283185307179586476925286766559;
+        radices_sum_prod = {0};
+        radices_prod     = {};
 
-        // Generate the table
-        size_t nt = twtc.size();
-        for(size_t i = 0; i < half_N; i++)
+        size_t prod      = 1;
+        size_t prod_next = radices.at(0);
+        size_t sum       = 0;
+
+        for(size_t i = 0; i < radices.size() - 1; ++i)
         {
-            double c = cos(TWO_PI * i / (2 * N));
-            double s = sin(TWO_PI * i / (2 * N));
+            auto radix      = radices.at(i);
+            auto radix_next = radices.at(i + 1);
 
-            twc_all[nt].x = c;
-            twc_all[nt].y = s;
-            nt++;
+            prod *= radix;
+            prod_next *= radix_next;
+            sum += prod * (radix_next - 1);
+            radices_sum_prod.push_back(sum);
+            radices_prod.push_back(prod_next);
         }
 
-        if(twts.alloc(total_length * sizeof(T)) != hipSuccess)
+        if(radices_prod.empty())
+            radices_prod.push_back(radices.at(0));
+
+        max_radix_prod = *(std::max_element(std::begin(radices_prod), std::end(radices_prod)));
+        min_radix      = *(std::min_element(std::begin(radices), std::end(radices)));
+
+        auto M          = radices.size() - 1;
+        auto last_radix = radices.at(M);
+        table_sz        = M ? radices_sum_prod.at(M - 1)
+                           + ((radices_prod.at(M - 1) / last_radix) - 1) * (last_radix - 1)
+                           + (last_radix - 1)
+                            : radices_sum_prod.at(0);
+
+        auto radices_bytes          = radices.size() * sizeof(size_t);
+        auto radices_prod_bytes     = radices_prod.size() * sizeof(size_t);
+        auto radices_sum_prod_bytes = radices_sum_prod.size() * sizeof(size_t);
+
+        if(radices_device.alloc(radices_bytes) != hipSuccess
+           || radices_prod_device.alloc(radices_prod_bytes) != hipSuccess
+           || radices_sum_prod_device.alloc(radices_sum_prod_bytes) != hipSuccess)
+            throw std::runtime_error("unable to allocate twiddle table kernel params");
+
+        if(hipMemcpy(radices_device.data(), radices.data(), radices_bytes, hipMemcpyHostToDevice)
+               != hipSuccess
+           || hipMemcpy(radices_prod_device.data(),
+                        radices_prod.data(),
+                        radices_prod_bytes,
+                        hipMemcpyHostToDevice)
+                  != hipSuccess
+           || hipMemcpy(radices_sum_prod_device.data(),
+                        radices_sum_prod.data(),
+                        radices_sum_prod_bytes,
+                        hipMemcpyHostToDevice)
+                  != hipSuccess)
+            throw std::runtime_error("unable to copy twiddle table kernel params");
+    }
+
+    void GenerateTable(const std::vector<size_t>& radices, hipStream_t& stream, gpubuf& output)
+    {
+        size_t           table_sz, maxElem, minElem;
+        gpubuf_t<size_t> radices_device, radices_prod_device, radices_sum_prod_device;
+
+        GetKernelParams(radices,
+                        radices_device,
+                        radices_prod_device,
+                        radices_sum_prod_device,
+                        maxElem,
+                        minElem,
+                        table_sz);
+
+        table_sz          = std::min(table_sz, length_limit);
+        auto total_length = attach_halfN ? table_sz + half_N : table_sz;
+
+        auto table_bytes = total_length * sizeof(T);
+        if(output.alloc(table_bytes) != hipSuccess)
             throw std::runtime_error("unable to allocate twiddle length "
                                      + std::to_string(total_length));
-        if(hipMemcpy(twts.data(), twc_all.data(), total_length * sizeof(T), hipMemcpyHostToDevice)
-           != hipSuccess)
-            throw std::runtime_error("failed to copy twiddle length "
+
+        if(table_bytes == 0)
+            return;
+
+        auto num_radices = radices.size();
+
+        auto blockSize  = TWIDDLES_THREADS;
+        auto numBlocksX = DivRoundingUp<size_t>(num_radices, blockSize);
+        auto numBlocksY = DivRoundingUp<size_t>(maxElem / minElem, blockSize);
+
+        auto device_data_ptr = static_cast<T*>(output.data());
+
+        hipLaunchKernelGGL(GenerateTwiddleTableKernel<T>,
+                           dim3(numBlocksX, numBlocksY),
+                           dim3(blockSize, blockSize),
+                           0, // sharedMemBytes
+                           stream,
+                           length_limit,
+                           num_radices,
+                           radices_device.data(),
+                           radices_prod_device.data(),
+                           radices_sum_prod_device.data(),
+                           device_data_ptr);
+
+        if(attach_halfN)
+        {
+            auto numBlocks_halfN = DivRoundingUp<size_t>(half_N, blockSize);
+
+            hipLaunchKernelGGL(GenerateHalfNTableKernel<T>,
+                               dim3(numBlocks_halfN),
+                               dim3(blockSize),
+                               0, // sharedMemBytes
+                               stream,
+                               half_N,
+                               N,
+                               device_data_ptr + table_sz);
+        }
+    }
+
+    void GenerateTable(hipStream_t& stream, gpubuf& output)
+    {
+        auto length       = std::min(N, length_limit);
+        auto total_length = attach_halfN ? length + half_N : length;
+        auto table_bytes  = total_length * sizeof(T);
+
+        if(output.alloc(table_bytes) != hipSuccess)
+            throw std::runtime_error("unable to allocate twiddle length "
                                      + std::to_string(total_length));
+
+        if(table_bytes == 0)
+            return;
+
+        auto blockSize = TWIDDLES_THREADS;
+        auto numBlocks = DivRoundingUp<size_t>(length, blockSize);
+
+        auto device_data_ptr = static_cast<T*>(output.data());
+
+        hipLaunchKernelGGL(GenerateTwiddleTableKernel<T>,
+                           dim3(numBlocks),
+                           dim3(blockSize),
+                           0, // sharedMemBytes
+                           stream,
+                           length_limit,
+                           N,
+                           device_data_ptr);
+
+        if(attach_halfN)
+        {
+            auto numBlocks_halfN = DivRoundingUp<size_t>(half_N, blockSize);
+
+            hipLaunchKernelGGL(GenerateHalfNTableKernel<T>,
+                               dim3(numBlocks_halfN),
+                               dim3(blockSize),
+                               0, // sharedMemBytes
+                               stream,
+                               half_N,
+                               N,
+                               device_data_ptr + length);
+        }
+    }
+
+public:
+    TwiddleTable(size_t _N, size_t _length_limit, bool _attach_halfN)
+        : N(_N)
+        , length_limit(_length_limit ? _length_limit : _N)
+        , attach_halfN(_attach_halfN)
+    {
+        half_N = attach_halfN ? (N + 1) / 2 : 0;
+    }
+
+    void GenerateTwiddleTable(const std::vector<size_t>& radices, hipStream_t& stream, gpubuf& twts)
+    {
+        auto use_radices = !radices.empty();
+        use_radices ? GenerateTable(radices, stream, twts) : GenerateTable(stream, twts);
+    }
+};
+
+template <typename T>
+class TwiddleTable2D : public TwiddleTable<T>
+{
+private:
+    size_t N1;
+    size_t N2;
+
+public:
+    TwiddleTable2D(size_t _N1, size_t _N2)
+        : TwiddleTable<T>(0, 0, false)
+        , N1(_N1)
+        , N2(_N2)
+    {
+    }
+
+    void GenerateTwiddleTable(const std::vector<size_t>& radices1,
+                              const std::vector<size_t>& radices2,
+                              hipStream_t&               stream,
+                              gpubuf&                    output)
+    {
+        if(radices1 == radices2)
+            N2 = 0;
+
+        auto blockSize = TWIDDLES_THREADS;
+
+        size_t           table_sz_1, maxElem_1, minElem_1;
+        gpubuf_t<size_t> radices_device_1, radices_prod_device_1, radices_sum_prod_device_1;
+
+        TwiddleTable<T>::GetKernelParams(radices1,
+                                         radices_device_1,
+                                         radices_prod_device_1,
+                                         radices_sum_prod_device_1,
+                                         maxElem_1,
+                                         minElem_1,
+                                         table_sz_1);
+
+        size_t           table_sz_2, maxElem_2, minElem_2;
+        gpubuf_t<size_t> radices_device_2, radices_prod_device_2, radices_sum_prod_device_2;
+
+        if(N2)
+            TwiddleTable<T>::GetKernelParams(radices2,
+                                             radices_device_2,
+                                             radices_prod_device_2,
+                                             radices_sum_prod_device_2,
+                                             maxElem_2,
+                                             minElem_2,
+                                             table_sz_2);
+        else
+            table_sz_2 = N2;
+
+        auto table_sz    = (table_sz_1 + table_sz_2);
+        auto table_bytes = table_sz * sizeof(T);
+        if(output.alloc(table_bytes) != hipSuccess)
+            throw std::runtime_error("unable to allocate twiddle length "
+                                     + std::to_string(table_sz));
+
+        auto device_data_ptr = static_cast<T*>(output.data());
+
+        auto num_radices_1 = radices1.size();
+
+        auto numBlocksX_1 = DivRoundingUp<size_t>(num_radices_1, blockSize);
+        auto numBlocksY_1 = DivRoundingUp<size_t>(maxElem_1 / minElem_1, blockSize);
+
+        hipLaunchKernelGGL(GenerateTwiddleTableKernel<T>,
+                           dim3(numBlocksX_1, numBlocksY_1),
+                           dim3(blockSize, blockSize),
+                           0, // sharedMemBytes
+                           stream,
+                           N1,
+                           num_radices_1,
+                           radices_device_1.data(),
+                           radices_prod_device_1.data(),
+                           radices_sum_prod_device_1.data(),
+                           device_data_ptr);
+
+        if(N2)
+        {
+            auto num_radices_2 = radices2.size();
+
+            auto numBlocksX_2 = DivRoundingUp<size_t>(num_radices_2, blockSize);
+            auto numBlocksY_2 = DivRoundingUp<size_t>(maxElem_2 / minElem_2, blockSize);
+
+            hipLaunchKernelGGL(GenerateTwiddleTableKernel<T>,
+                               dim3(numBlocksX_2, numBlocksY_2),
+                               dim3(blockSize, blockSize),
+                               0, // sharedMemBytes
+                               stream,
+                               N2,
+                               num_radices_2,
+                               radices_device_2.data(),
+                               radices_prod_device_2.data(),
+                               radices_sum_prod_device_2.data(),
+                               device_data_ptr + table_sz_1);
+        }
     }
 };
 
@@ -146,6 +334,7 @@ public:
 template <typename T>
 class TwiddleTableLarge
 {
+private:
     size_t N; // length
     size_t largeTwdBase;
     size_t X, Y;
@@ -161,34 +350,33 @@ public:
         tableSize = X * Y;
     }
 
-    std::vector<T> GenerateTwiddleTable()
+    void GenerateTwiddleTable(hipStream_t& stream, gpubuf& output)
     {
-        std::vector<T> wc(tableSize); // cosine, sine arrays
-        const double   TWO_PI = -6.283185307179586476925286766559;
+        auto table_bytes = tableSize * sizeof(T);
+        if(output.alloc(table_bytes) != hipSuccess)
+            throw std::runtime_error("unable to allocate twiddle length "
+                                     + std::to_string(tableSize));
 
-        // Generate the table
-        size_t nt  = 0;
+        if(table_bytes == 0)
+            return;
+
+        auto blockSize = TWIDDLES_THREADS;
+
         double phi = TWO_PI / double(N);
-        for(size_t iY = 0; iY < Y; ++iY)
-        {
-            size_t i = static_cast<size_t>(1) << (iY * largeTwdBase);
-            for(size_t iX = 0; iX < X; ++iX)
-            {
-                size_t j = i * iX;
 
-                double c = cos(phi * j);
-                double s = sin(phi * j);
+        auto numBlocksX = DivRoundingUp<size_t>(X, blockSize);
+        auto numBlocksY = DivRoundingUp<size_t>(Y, blockSize);
 
-                // if (fabs(c) < 1.0E-12)	c = 0.0;
-                // if (fabs(s) < 1.0E-12)	s = 0.0;
-
-                wc[nt].x = c;
-                wc[nt].y = s;
-                nt++;
-            }
-        } // end of for
-
-        return wc;
+        hipLaunchKernelGGL(GenerateTwiddleTableLargeKernel<T>,
+                           dim3(numBlocksX, numBlocksY),
+                           dim3(blockSize, blockSize),
+                           0, // sharedMemBytes
+                           stream,
+                           phi,
+                           largeTwdBase,
+                           X,
+                           Y,
+                           static_cast<T*>(output.data()));
     }
 };
 
@@ -202,34 +390,16 @@ gpubuf twiddles_create_pr(size_t                     N,
     if(largeTwdBase && length_limit)
         throw std::runtime_error("length-limited large twiddles are not supported");
 
-    gpubuf         twts; // device side
-    std::vector<T> twtc; // host side
+    gpubuf      twts;
+    hipStream_t stream;
+
+    if(hipStreamCreate(&stream) != hipSuccess)
+        throw std::runtime_error("hipStreamCreate failure");
+
     if((N <= LARGE_TWIDDLE_THRESHOLD) && largeTwdBase == 0)
     {
-        TwiddleTable<T> twTable(N, length_limit);
-        if(radices.empty())
-        {
-            twtc = twTable.GenerateTwiddleTable();
-        }
-        else
-        {
-            twtc = twTable.GenerateTwiddleTable(radices); // calculate twiddles on host side
-        }
-
-        if(attach_halfN)
-        {
-            twTable.AttachHalfNTable(twtc, twts);
-        }
-        else
-        {
-            if(twts.alloc(twtc.size() * sizeof(T)) != hipSuccess)
-                throw std::runtime_error("unable to allocate twiddle length "
-                                         + std::to_string(twtc.size()));
-            if(hipMemcpy(twts.data(), twtc.data(), twtc.size() * sizeof(T), hipMemcpyHostToDevice)
-               != hipSuccess)
-                throw std::runtime_error("failed to copy twiddle length "
-                                         + std::to_string(twtc.size()));
-        }
+        TwiddleTable<T> twTable(N, length_limit, attach_halfN);
+        twTable.GenerateTwiddleTable(radices, stream, twts);
     }
     else
     {
@@ -237,31 +407,18 @@ gpubuf twiddles_create_pr(size_t                     N,
 
         if(largeTwdBase == 0)
         {
-            TwiddleTable<T> twTable(N, length_limit);
-            twtc = twTable.GenerateTwiddleTable();
-            if(twts.alloc(twtc.size() * sizeof(T)) != hipSuccess)
-                throw std::runtime_error("unable to allocate twiddle length "
-                                         + std::to_string(twtc.size()));
-            if(hipMemcpy(twts.data(), twtc.data(), twtc.size() * sizeof(T), hipMemcpyHostToDevice)
-               != hipSuccess)
-                throw std::runtime_error("failed to copy twiddle length "
-                                         + std::to_string(twtc.size()));
+            TwiddleTable<T> twTable(N, length_limit, attach_halfN);
+            twTable.GenerateTwiddleTable(radices, stream, twts);
         }
         else
         {
             TwiddleTableLarge<T> twTable(N, largeTwdBase); // does not generate radices
-
-            twtc = twTable.GenerateTwiddleTable(); // calculate twiddles on host side
-
-            if(twts.alloc(twtc.size() * sizeof(T)) != hipSuccess)
-                throw std::runtime_error("unable to allocate twiddle length "
-                                         + std::to_string(twtc.size()));
-            if(hipMemcpy(twts.data(), twtc.data(), twtc.size() * sizeof(T), hipMemcpyHostToDevice)
-               != hipSuccess)
-                throw std::runtime_error("failed to copy twiddle length "
-                                         + std::to_string(twtc.size()));
+            twTable.GenerateTwiddleTable(stream, twts);
         }
     }
+
+    if(hipStreamSynchronize(stream) != hipSuccess || hipStreamDestroy(stream) != hipSuccess)
+        throw std::runtime_error("hipStream failure");
 
     return twts;
 }
@@ -298,35 +455,19 @@ gpubuf twiddles_create_2D_pr(size_t N1, size_t N2, rocfft_precision precision)
     }
     radices1.insert(radices1.cbegin(), kernel.factors.cbegin(), kernel.factors.cbegin() + count);
     radices2.insert(radices2.cbegin(), kernel.factors.cbegin() + count, kernel.factors.cend());
-    if(radices1 == radices2)
-    {
-        N2 = 0;
-    }
 
-    TwiddleTable<T> twTable1(N1, 0);
-    TwiddleTable<T> twTable2(N2, 0);
+    gpubuf      twts;
+    hipStream_t stream;
 
-    auto           twtc1 = twTable1.GenerateTwiddleTable(radices1);
-    std::vector<T> twtc2;
-    if(N2)
-    {
-        twtc2 = twTable2.GenerateTwiddleTable(radices2);
-    }
+    if(hipStreamCreate(&stream) != hipSuccess)
+        throw std::runtime_error("hipStreamCreate failure");
 
-    // glue those two twiddle tables together in one malloc that we
-    // give to the kernel
-    gpubuf twts;
-    if(twts.alloc((twtc1.size() + twtc2.size()) * sizeof(T)) != hipSuccess)
-        return twts;
-    auto twts_ptr = static_cast<T*>(twts.data());
-    if(hipMemcpy(twts_ptr, twtc1.data(), twtc1.size() * sizeof(T), hipMemcpyHostToDevice)
-           != hipSuccess
-       || hipMemcpy(twts_ptr + twtc1.size(),
-                    twtc2.data(),
-                    twtc2.size() * sizeof(T),
-                    hipMemcpyHostToDevice)
-              != hipSuccess)
-        twts.free();
+    TwiddleTable2D<T> twTable(N1, N2);
+    twTable.GenerateTwiddleTable(radices1, radices2, stream, twts);
+
+    if(hipStreamSynchronize(stream) != hipSuccess || hipStreamDestroy(stream) != hipSuccess)
+        throw std::runtime_error("hipStream failure");
+
     return twts;
 }